Create Signature-dev-generic.html
This is the "starting" point for the final html that will be inserted into the legal disclaimer.

This is the base HTML used to generate a sample signature.  The code is formatted nicely for ease of reading and editing.

NOTE:  this doesn't have active directory attributes added, nor does it contain inline-css which is needed for proper exchange and outlook handling.